Refactoring and linting
