+++ server/src/ga/GaWorker.ts
