+++ server/src/index.ts
@@ -1,6 +1,6 @@
 require("dotenv").config();
 
-import app from "./app/app";
+import app from "./app/App";
 
 // PORT env variable is provided by Heroku
 const port = process.env.PORT || 9000;
